.emacs.el: move some code around
The code added in the last two commits depends
on my lambda-let and with-advice helpers, so I
am moving it below that.

Not really any good place to put it, since it's
no longer pure simple config like the original
use-short-answers boolean setting was. In some
ways this is motivated by using Evil, and my
my minibuffer coloring to indicate vi state,
and were that the sole reason, I would put it
inside Evil's config. But this makes sense
independently of Evil too.