.emacs.el: smoother kill buffer UX on " k"
* " k" now kills current buffer,
* universal-prefixed " k" (such as "  k")
  prompts to kill any other buffer.

The vast majority of my uses of kill buffer
are to kill the current buffer. And when I
do want to kill another buffer, it's usually
multiple times in a row to clean up a few
buffers (too few to open Ibuffer) and it's
annoying to have the current buffer in the
candidate list.