.emacs.el: fix space-prefixed universal argument
Subtle bug: all of the universal argument
functions set the transient keymap, which
ends up in effect for the next key sequence
read of Emacs' command loop (transient key
map is removed in the pre-command-hook).
But `command-execute` does not completely
immitate a full iteration of Emacs' input:
in particular, it does not run post- and
pre- command hooks (it also doesn't set
`{last,this}-command`, apparently).

The fuller fix would be to properly do
those steps, but here's a quick fix: unset
the transient map by calling the function
which is returned by `set-transient-map`,
which is (undocumented, but probably
intentionally) returned by the
`{universal,digit,negative}-argument`
functions.