.emacs.el: make Enter count as "yes" in `y-or-n-p`
I'm pretty sure this has annoyed me in the past.
Space counts as "y", Backspace counts as "n",
and thanks to Evil in minibuffers, Escape first
exists to vi normal state and then counts as a
quit (an even stronger "no" which aborts the
whole operation rather than letting it proceed
on a "no" answer, but in most cases this is the
same thing a "no" answer would do). Enter is an
obvious and natural "yes", more than Space.