.emacs.el: add confirm-p function
Just a basic confirmation which only
accepts Enter as a yes.

This is what I was trying to do three
commits ago with `read-multiple-choice`
with just return as the choice, without
any of the problems caused by the
limitations of `read-multiple-choice`
being so character-oriented.

The advantage of this over something
like `y-or-n-p` is that it's a safer
prompt in situations where the user may
be already pressing other characters,
due to pre-planned movements or muscle
memory or because they got to this
prompt by a misclick/typo of another key
(where that other key would've been
rightly followed by other keystrokes).