.emacs.el: fix/improve smoother-kill-buffer prompt
Switching from `read-multiple-choice` to
`y-or-n-p` introduced the problem that I
was originally thinking of when I chose
to use `read-multiple-choice` with only
one Enter choice instead of `y-or-n-p`:

Let's say I hit " l" for `consult-line`.
Naturally I am already typing my intended
search query by the time I realize that
oops I hit " k" on accident instead.

It just so happens that literally after
making the commit to switch to `y-or-n-p`,
I made that exact mistake, and my search
was for "y-or-n-p" so I immediately killed
the current buffer instead of searching it.