.emacs.el: fix confirm-p for terminal
Apparently in this case in particular, unlike
any others, *both* carriage return (Ctrl+m)
and the "return" symbol need to be bound. GUI
Emacs needs the latter, terminal Emacs needs
the former.

(The common case is that just carriage return
needs to be bound, and it works for both. I
presume this is because `read-with-minibuffer`
is more raw, using just the keymap given and
not any implicit translation maps (or maybe
my default binding shadows something?).)