.emacs.el: better coloring for Denoted file names in Dired
This wrangles font-lock to fontify Denoted names
much more usefully. In particular, it allows me
to fontify my standard delimiters between words,
which massively improves readability, which is an
especially huge optimization to my todo system,
where I actually look at the task files in Dired
as my task list and so on. I also use the Denoted
naming convention (with my additions of optional
datetime and allowing `--`) for other files - for
all of my music files, starting to use it for
various documents, converting my pictures and
videos to it, and so on.

Not sure yet about any specific color choices,
but I think I want something vibrant enough to
read with full sun on the screen (like many of
my colors, especially the 16 standard terminal
foreground colors) and with enough difference
that it's instant to visually pick apart each
separate file name. (I am mostly happy with my
choice of dim gray for delimiters and extensions,
despite the lack of sun visibility, precisely
because they're made to fade into the background,
and I can bind a command to toggle the colors
off (maybe I even did and just forgot) for the
few times I want to see them more clearly.) (I
and pretty happy with my not-too-prominent band
of color over the month and minute, but those
color choices still need bright sunlight testing.)
(I am pretty sure I am dissatisfied with other
colors at least a little, because the defaults
for title, tags (keywords), and prefix (signature)
have a way of making the the title the least
prominent and dimmest of the three, and harder
to read because of it.)

A roughly solid 18-hour effort, maybe closer to
24, plus an hour or two of initial scouting a
couple/few days ago.