.emacs: remove " sg" for searching tasks
Literally forgot I had this. Funny, given
that I use " sh" for task tag search and
basically every time dimly remember that
there was the regular task search on " sg".

I might re-add this later, probably with
actual code sharing between note search
and task search by then. It amuses me that
I improved note search to quit the notes
buffer, but apparently didn't think to do
that with the task buffer (or I thought of
it and decided against it, in itself funny
and similarly indicative that I was using
it so little that it faded from my mind).

I suspect I'll see more reasons for search
on tasks as I use tasks more heavily.