Revert ".emacs.el: smoother-kill-buffer: even smoother"
This reverts commit 309ad901400079cca1e3c5f2604068c1766f4614.

This is a great idea, but vulnerable to the typo
problem: for unsaved buffers, it brings back the
risk of killing a buffer by for example hitting
" l" (with "l" typoed as "k") to do a search
which starts with "y" (or just contains "y" in
the beginning if I type fast enough, or anywhere
if I type long enough without looking and without
sound/vibration cues).

At some point I may rethink all of my keybinds
to arrange functionality so that this problem
categorically can't happen, and then I can
bring this back.