.emacs.el: combine `note` and `evil-note-paste`
" m" (the main note keybind) with vi register
or prefix argument will now paste. (Prefix
argument value is not used - any prefix value
is the same as pasting without specifying the
register; there is some room here for using
numeric prefixes to select from the kill ring
or something, and that would be more ergonomic
than numbered vi registers thanks to my
space-prefixed universal argument keybinds, but
eh, I'll add it later if I miss it enough.)

Not entirely sure if this is the play, but it
makes a decent amount of sense to me. It frees
up "n" for something else again (or "m" if I
move notes to "n"), and the main note keybind
wasn't doing anything useful with either the
prefix argument or the vi register anyway.