.emacs.el: add consult-line-{next,previous}
One of the things I have often wanted is to
repeat a Consult line search same as I would
a vi search: hit "n" or "N" to go to the
next or previous search result.

This implements the actual next and previous
functions for Consult's line search. It's
hacky and in particular there are race
conditions:

1. the UI might flicker because for a moment
   the minibuffer is opened, and with my
   dimming of inactive windows, the whole
   buffer might flicker in color;

2. sufficiently fast keypresses can end up
   end up going to the search bar, which at
   best swallows the key (Space, since this
   doesn't effect Orderless search in most
   cases) or at worst breaks the search
   (for example, if I had this bound on "n"
   and hit "n" too fast, it would add an "n"
   to the end of the search string).