.emacs.el: add defer-input macro
`defer-input` allows wrapping a piece of code so
that all user input handling is deferred until
after the code finishes executing:

    (defer-input
        ...)

Inside the body of `defer-input`, the keymap
`defer-input-map` can be used to allow specific
keys or events to still be handled.

Since events can be arbitrary symbols, this also
means that `defer-input` is compatible with hacks
that inject values into `unread-command-events`.