.emacs.el: bind Ctrl+G in consult-line-next
When things are working correctly, this does
nothing. But if for any reason there's ever
a bug or anything during consult-line-next
or consult-line-previous that causes the
command to get stuck in the minibuffer (such
as the current bug when running them in a
buffer that doesn't have any matches for the
last consult-line search), this provides an
escape hatch instead of locking up Emacs.