.emacs.el: integrate vi and consult-line searches
"n" and "N" now cycle the last of either vi
search or Consult's line search. I've wanted
and intuitively reflexively tried to do/use
this ever since I got used to using Consult's
search at least a year ago.

The error message for "n" and "N" when there
hasn't been any recent search is now cleaner
as well. (Evil's native search errors out in
a way that needlessly quotes the error reason
as if it was a search query the user typed.)