.emacs.el: fix no-extension denoted renames
`denoted-extension-get` now returns leading ".".
This simplifies usage. I was on the fense at
the time and only just barely tipped in favor
of returning it without the leading period.
The fact that it helps avoid a bug with less
code is good enough to tip that precarious
balance the other way.