.emacs.el: task-list: fix marking top line
Hack on top of hack: Dired keeps track of which
lines show directories by using the position of
the start of each line, so when I delete the top
line instead of narrowing, it ends up believing
that the new top line (a file) is also a
directory still, and that takes precedence over
it being a file for marking purposes.

This fixes that by letting Dired know that the
the directory line which previously started at
character one now starts at character zero
(which does not exist, and this prevents it
from ever matching, and is fairly semantically
accurate and precise).