.emacs.el: Denoted: disallow "-" in prefixes
I was on the fence about allowing "-" in prefixes.
I have no use for it but it seemed possibly useful.
But Denote logic doesn't currently support "-" in
signatures, which is just enough of a downside to
tip the balance against it for now (so that I
can't use my renaming functions to create names
that those renaming functions then mishandle).