.emacs.el: Denoted: change prefix to suffix
Per my last commit message, this is better for
the only usecase I have for Denoted signatures
at this time, and would probably be better for
several of the other likely usecases as well.

(The only situation I see where I would miss
having the prefix before the title is if I had
multiple pieces of writing in a series, and I
didn't want to have an overloaded datetime to
sort them, and putting each one in its own
folder was unappealing. That's a possible use
case, but one I don't yet find myself wanting.
Maybe if I was authoring books or a blog post
series, and for some reason I didn't want to
have the chapter/post number in the title, but
did want them all sorted in the same folder.
Seems... plausible, not unrealistic, but I
suspect in those cases I'd be happy to solve
this in the tooling I'd be already using for
that use case.)

This is also more natural, because if I make
it a suffix instead of a prefix, then the case
of both title and affix looks like the nicer

    my-title==AFFIX.md

instead of

    ==AFFIX--my-title.md