.emacs.el: workaround bad hexl edge case
Hexl's cursor behavior has a bad inconsistency
where inserts move the cursor forward unless
both

1. the cursor is on the last byte in the buffer,
   and

2. the last byte in the buffer just happens to
   line up on the right edge of hexl's columns.

Arguable if this is intentional or good, but if
so, then the second point is a bug, as is the way
it combines with multibyte insert to silently do
the wrong thing (for example, inserting "ф" would
normally insert 0xD1 0x84, but with the current
code, inserting it at the very end of the buffer
silently inserts 0x84, instead of the obviously
righter behavior of inserting 0xD1 and signaling
an error).

But to me, having the cursor behave differently
at the end of the buffer is bad design, because
it forces any other code composition to know
about and work around that edge case (the bug
with multibyte insert is literally an example of
this principle in action - the code would do the
right thing for free if the cursor always moved
forward on insert; since the cursor doesn't, the
author of that code needed to proactively invent
that edge case, test it, and write extra code;
since the author(s) didn't, which is only natural,
we have this bug).

So this commit injects a fix into hexl to make
point always advance.