.emacs.el: work around Denote's new no-confirm+save
Denote 2.3.0 overloaded the no-confirm parameter
of `denote-rewrite-front-matter` to both stifle
the y-or-n prompt and to unconditionally save
the file. Personally I never want either of

1. unconditionally save, or

2. ask to confirm

so this is in some ways worse since I have to go
back to advising either y-or-n-p (or save-buffer)
to independently control these behaviors.

My wrapper could become a little cleaner, but I
don't want to rely on the meaning of no-confirm
staying overloaded like this (and in fact I am
filing a bug report arguing that it shouldn't).