.emacs.el: add fixed-hexl-save-buffer
This makes saving files in hexl mode actually
usable+good, so that it neither discards undo
history nor loses your place in the buffer.

This is done by dehexlifying the contents in
another buffer and writing out those contents
instead of dehexlifying in the actual file's
buffer. Another possibility would be to just
dehexlify and rehexlify in-place and just
keep those changes from being added to the
undo history, but this was quicker and easier,
with zero need to fight the implementation in
hexl itself.

This might seem inefficient, but Hexl already
does the same thing, it just chooses to skip
all the benefits by using the temp buffer as
a way to save the hexlified contents to not
have to rehexlify. This has some tiny dubious
benefits like reusing a more of the default
save-buffer machinery, including precious flag
handling, save hooks, and so on. If this is
important, it could be reimplemented in a few
ways, but none of those features are worth the
fundamental, essential editor functionality
of undo history and not jumping you to the end
of the file after every save. As I need those
other features, I can add them.