.emacs.el: optimize Denoted coloring in Dired
This replaces my heavy-weight fontification
(which needed four separate top-level passes,
each with its own custom matcher function)
passes with a lighter-weight version which
just uses one pass and seems (at a cursory
glance) to work for all of my file names.

This new version doesn't handle the full
flexibility in file names that the previous
version would handle, but I'm fine with that
since it brings the latency back down to a
tolerable level when lots of files are in
view (62, on my biggest screen), and I don't
foresee using that extra flexibility, let
alone in ways that I couldn't incrementally
add into this version.

This reverts commit 462281ba74a0c9cedb3da770f6d4abe5dc5e7c00.