.emacs.el: eshell+histdir: more conventional Up/Down
Seems that this experminent comes to an end: I
reach for Up/Down while in scrollback to cycle
history regularly, but I don't normally find
myself accidentally using arrow keys while in
insert mode in scrollback to try to move in
the scrollback (that I can remember). It's too
useful and familiar, and consistent with how
things work in my REPL wrappers, to lose just
to protect against accidental use of arrows in
vi insert state in scrollback (and if/when that
does become an issue, I'd now prefer to tackle
that how I tackle any other unintented jumps...
so maybe I'd make Up/Down mid-scrollback add an
entry to vi jump lists and go from there, etc).