.emacs.el: with-editor: make kill-buffer work
For some reason, with-editor makes the decision
to suppress kill-buffer instead of just making
it equivalent to cancel. It's super weird to me
but okay, I'm sure in some contexts (like with
how it couples with Magit) it makes sense. But
in the raw standalone usage of with-editor,
cancel replaces (and enhances/supplements)
killing the buffer, and I expect such things to
not try to be smart about *how* I'm killing the
buffer - if you're going to remap kill-buffer,
ideally programmatic calls to kill buffer should
do the same thing.