histdir-repl: fix/optimize fast repeat of x/X/Backspace/Delete
When holding down or rapidly mashing keys that
delete a single character, "entering into" the
input, empirically re-finding the start and end
of input, and sending a bunch of cursor movement
keystrokes is profoundly inefficient and slow
enough that it causes errors (deletions end
up happening in the wrong place because (to
a cursory interpretation), waiting for the
terminal to settle after sending cursor-moving
inputs allows processing the next keystroke,
which then moves the cursor again, and then
waits finish and send character-deleting inputs
with the cursor not where they expected).

Using my `defer-input` eliminates the input race
and interleave, but also combines with the tiny
sleeps (which are needed tobwait for the program
to update the terminal) causes the UX to slow
down or even freeze momentarily, so for example
you hold down Backspace at the end of "abcdefgh"
for a moment, the UI freezes and a split second
later you're down to "ab", without the normal
smooth deletion of every character individually.

A simple optimization which also almost entirely
eliminates the need for my `defer-input` is to
special case the deletes when there's no numeric
prefix argument and Emacs point is already in
the same place as the input: just send the input
for Backspace/Delete to the terminal, no need to
"enter" input and do all the fancy inputs. So
that's what this commit does, and I'm saving the
`defer-input` until I actually see issues without
it (logically it seems possible, but I want to
guage likelihood and reduce risk of knock-on
issues while moving on to higher priorities).