.emacs.el: whitespace tweak/fixup
Apparently when I first wrote this use
of `with-hook`, I wanted four spaces
from the `'` and not from the `(`, as
I do in calls to my `with-advice` and
so on.

But looking at it now, there's no good
reason for that, since unlike my other
macros (`with-advice`, etc), `with-hook`
has a nested parenthesis since `add-hook`
is variadic, and there's no real benefit
to breaking from a more typical lisp-y
indentation style (but there is at least
one small but frequent downside - it's
not how the built-in Tab behavior does
it).