.emacs.el: double-tab: more flexible indenting
This allows me the best of both worlds between
`lisp-indent-offset` being `nil` (fancy smart
Elisp indenting) and `4` (much dumber, simpler,
but therefore more controllable indenting).

I mainly want this when dealing with certain
nested constructs like the value form of a
line-wrapped name-value pair in a `let` form.

I find

    (let ((some-long-name
              (a-long-function-call ...))
          (other-name (short a b)))
        ...)

much more readable than the default

    (let ((some-long-name
           (a-long-function-call ...))
          (other-name (short a b)))
        ...)

because that extra indent makes it much more
visually obvious and easy-to-notice that the
second line is somehow "part of" or "within"
the last clause in the prior line.

So now I can get default smart indent behavior
with one Tab, and if that's not what I want I
can force the dumber behavior (and it's still
transparent to Tab triggering completion).