.emacs.el: with-nested-command-state helpers
Lots of commands technically are done running
once they've set up some interactive UX, but
conceptually they aren't done running until
the user is done with that interactive UX.

For example, `consult-line` is technically
done once the minibuffer with candidates is
raised, but conceptually it isn't done until
after the user selects a search candidate or
quits the search.

More relevantly, functions which use the
minibuffer to ask for confirmation break
encapsulation: if a command uses `y-or-n-p`,
my "fixed" `read-multiple-choice`, or my
`confirm-p`, the user response gotten with
`read-from-minibuffer` counts as another
command, which means that by calling either
of those functions we wrongly mutate the
last and current command information unless
we protect those variables.

It's always correct for commands which nest
other commands to preserve `last-command`
and `this-command` (and always wrong not to).
Probably it's also appropriate to preserve
`real-last-command` and `real-this-command`.
`y-or-n-p` does the latter, for example. In
fact, arguably not preserving it breaks
expectations/invariants of those two variables
not being changed by code than letting them
get changed by nested commands does.

So this commit implements a couple helpers
which protect all of those variables, and we'll
see how it goes. The first helper is a block
macro, while the second is a function which
can be used as "around" advice (and incidentally
in all the places `funcall` could be used).