.emacs.el: use nested command state for y-or-n-p, read-multiple-choice, and confirm-p
Right now these functions are all slightly
inconsistent. This fixes that while reducing
the `[real-]{this,last}-command` boiletplate
in my code (while clearing up the big-picture
intent of that boilerplate), and "fixes" them
to properly encapsulate the implementation
detail that replying to them is a command.

In the long run, I probably want to do this
for `completing-read` and other functions that
generally are called by commands which aren't
done yet. But let's see how this goes.