.emacs.el: tumblr--ensure-file: fix parenthesis
Before the indentation fix in the last commit,
the indentation here didn't match the
parentheses.

Based on what I know of myself and how I think,
I am confident I intended to run the hook only
in the success case of the whole function (as
the indentation suggests), rather than running
it in the success case of the cleanup of the
temporary file.

The ideal behavior would probably be to have it
in a second unwind protect clause so that even
if deleting the file or refreshing modified state
fails, the hook still gets called.

However, I want it in just the success case
because the failure mode is harmless and
usefully educational - I want to understand
why the hook needs to be called, and at least
emprically confirming if it ever ends up being
needed in the error case would be clarifying.