.emacs.el: fix my vc-dir disabling
The previous version seems to not take
effect - I guess the function is loaded
or the definition is evaluated (again?)
after my config code. Weird but whatever,
advice seems to work at a first glance.

I've also figured out how `vc-dir` is
getting invoked: in Termux, a touch on
the vc part of the modeline triggers
the vc-dir entry of that menu (on a
normal desktop environment, a left mouse
click on the vc modeline part would only
open that menu without choosing an action).
I haven't figured out the deeper how of
why a tap has that effect.