.emacs.el: task-schedule-repetition: deplete to R0
Instead of removing the repetition suffix entirely,
leave it with the number of repetitions set to zero.

This enables things like cycling forward temporarily,
just to check what the resulting date is or for
testing, and then cycling back.

I also suspect it might be useful for me to look
at a recurring task file on its last occurrence
and know at a glance that it was a recurring task.

(I probably will also eventually want to store
something in the front matter to indicate when
the first occurrence was - the total number of
repetitions would be sufficient, but not very
human-friendly. But I don't yet know if I'll
end up using this enough to get to that point.)