.emacs.el: make toggles just show state at first
I find myself usually hitting my toggle function
keybinds twice just because I'm not sure if that
toggle is already in the state I want (it is,
almost every single time).

I had several ideas for how to improve this, for
example using presence/absence of prefix argument
to control if it toggles or just displays state,
but for now I think just displaying on first
consecutive call and only toggling if the command
is called more than once in a row makes a lot of
sense. Enough to try but I suspect I'll like it
indefinitely.