.emacs.el: add += helper macro
Allows simplifying

    (setq variable (+ variable something))

to

    (+= variable something)

and the more general

    (setf place (+ place something))

to

    (+= place something)

This makes code significantly more readable
some cases, by improving signal-to-noise.

Holding off on adding `-=` and the like
until I have more need for them. In the
bigger picture, there's clearly also value
here from generalizing one step up: a macro
that combines any operation on a value or
place with `setf`.