.emacs.el: decoded-time-days-until-weekday
Helper function to compute the number of
days to get to a given day of the week
from a given decoded time.

Return values are in the range [0, 6]
(as opposed to [1, 7]). My intuition is
that this is the right design choice in
general, but I'm definitely biased by my
current use-case for datetime parsing.