.emacs.el: datetime-parse: day-of-week trumps integer
Before: `1 tue` has the 1 bind as the year.

After: `1 tue` has the 1 bind as the hour.

This seems more useful to me, is more consistent
with how some of the other stuff works (`5 feb`
has the 5 bind as the day), and when I came back
to this after a couple months of not specifying
much with day-of-week, my intuition was that it
would work as implemented in this commit.

This does have the downside that `2022 1 tue`
has the 1 also bind as the hour, even though I
find myself thinking that I'd normally want it
to bind to the month in that case.