.emacs.el: window-state: swap "gp" and "gP"
It turns out that usually (feels like always),
when I am doing "fast paste" from one split to
another, I'm intending to work in the new copy,
while leaving the old copy where it is, kinda
like a call stack/tree - the new copy of the
view into the buffer is where the new idea is
is acted on.

Meanwhile, I'm leaving "x" and "X" the way
that they are, because when swapping, it's
usually (always?) because I want to work on
that other window split's contents (in the
current split) rather than wanting to work
on the current contents (in the other split).

I'm not sure yet if I would consider this
the right default in a public release.