.emacs.el: datetime-parse: simplify bad word error
Instead of raising an error for all bad words,
just raise an error with the first bad word.

In some ways, this is just strictly worse, hence
why I initially went out of my way to signal an
error containing all bad words at once.

But the one benefit is that the first error seen
is always the error being signaled. So when you
type "foo", you get the error

    Bad word: "foo"

and when you type "foo 50hours +5hours" you
still get

    Bad word: "foo"

even if (as will happen in my upcoming changes)
the code starts to report errors on invalid
values during datetime arithmetic.

(The other sensible alternative is to try to
aggregate all errors, but that's a little too
intractable, because each invalid datetime
part "changes" the interpretation of the next
parts by being invalid - if it was replaced by
a valid value instead, the next ambiguous
integer would mean something else.)