.emacs.el: fixed-decoded-time-add: fix leap day correction
This fixes a bug introduced in commit
7f0615ba08ad96881fe3c68b45372cb90021a05b.

Adding 1 year to a February 29th should land on
February 28th (or arguably Marh 1st) of the next
year, but the change in that commit prevents
that normalization from happening unless there's
also a day offset.

This adds an explicit separate correction, not
part of the normal arithmetic, which runs only
when a year and/or month delta results in a
February 29th, before performing any of the
lower-magnitude/higher-precision arithmetic.