.emacs.el: add `any` and `all` helper functions
Just like the ones from Python: take an iterable
and return true if any/all of the values are
non-nil.

This description doesn't cover the nil case where
sequence has no elements, but the behavior is the
fairly reasonable: `any` will return `nil` for an
empty sequence, while `all` will return `t`; this
naturally follows from the simplest implementation,
and sorta makes sense because the answer to "are
any of these elements true?" is "no" where there
are no elements, while "are all of these elements
true?" is "yes" when there are no elements.

For now it just uses `dolist`, which I think
will fail to work for most/all iterables that
aren't lisp lists, but I'll expand it as needed.