.emacs.el: fixed-decoded-time-add: fix nil checks
Another partially-defined datetime corner case:
if the month is nil but the day delta isn't, or
the day is nil but the month delta isn't, the
special handling for leap days breaks because
it uses `=`, which requires its arguments to be
either numbers or markers.

This commit switches the `=` to `equal`, as I
normally do. (The only reasons for using `=`
here were consistency with other numeric
comparisons in `fixed-decoded-time-add`.)