.emacs.el: datetime-read preview: show bindings
Visually indicate which integers were bound to
which part of the datetime. (Unfortunately,
this is purely based on text faces, namely
color and bold weight. In principle this is bad
accessibility and I really don't like that, but
in practice it's good enough for me and most
people in most scenarios.)

The implementation is an awful mess. I dislike
the deep coupling between parse and read preview
(although it seems rather intrinsic to the goal),
I dislike the mutation of `words`, and the naming
isn't particularly clear anymore.... but I can't
keep spending time on this. Something that works
is crucial right now in my life - making the code
nicer can wait... must wait.

Bindings are not shown if they happen earlier in
the parse, while still parsing other words. The
motivation for this was consistency in the cases
where integers get bound but are then overwritten
by later words.