.emacs.el: refactor datetime-parse--1 into macro
As nice as it was to Properly make it a pure
function (although it became a little less pure
once `bindings` was involved, plus it was never
fully pure since `parsed`, `previous` and
`integers` were often mutated despite being
returned as well), it's currently slowing me
down more than it's helping, since the whole
loop body only grows ever more complected with
the overall parse (in order to keep enhancing
the preview feature).

I think the True Solution is to gather up all
of the mutable state that only exists for the
purpose of preview generation, pass it in as
one object, and return it as one object (monad
vibes). That would allow for a cleaner interface
that's still explicit, functional/pure, and
even better at being self-explanatory. But for
now, Gotta Go Fast.