.emacs.el: datetime-parse: fix future bias
The datetime-parse v2 change regressed the
future biasing logic to ignore lower parts
of the datetime. For example, the input

    may 1st

on May 2nd should parse as the next year's
May 1st. As of the v2 change, the future
bias logic was picking the current year,
since May was still available now slash in
the immediate future.

This commit fixes that. It also simplifies
the logic significantly (and in principle,
significantly optimizes it too, though I
didn't actually measure): instead of each
part of the datetime being compared in a
separate operation, the entire datetime is
xor'ed into a single integer (which will
fit within 64 bits for all realistic years)
and then that integer representation is
compared in a single operation.