.emacs.el: datetime-read preview: better diff-or-not
Instead of ignoring all ambiguous integers for count
purposes, only ignore the ones that are followed by
at least one word that's not an ambiguous integer.

This way:

* Something like `5 5` will show future bias diff
  (or not) exactly the same as `may 5` would.

* Something like `1 30 sunday` still doesn't
  show the diff (since in that situation the
  diff is the result of the bindings changing,
  which is shown by the colors, while the diff
  itself is unhelpful noise or even
  confusing/misleading).

There's still a problem with inputs like `5 5th`,
which end up not showing a diff even though the
behavior is otherwise exactly the same as `5 5`,
`may 5`, and `may 5th`.