.emacs.el: " vC" now supports prompt for commit-ish
Because the few seconds of my time that it would take
to go to a shell, even though that happens only once
every few months, is worth it, apparently.

I was actually surprised it didn't already do this
because I am so used to all of my other Git keybinds
which take numeric prefix arguments to select commits
relative to HEAD having this capability.