.emacs.el: datetime-parse: validation consistency
Make an ambiguous integer like `50` bypass
validation the same way that `50d` does.

(Arithmetic on that part of the datetime
will still error out as normal.)

This allows something like `5 45 su` (for
"5:45 on Sunday") to work the same way
that `5 30 su` does: if ambiguous integer
binding validates, then the `45` would
error out when computing the bindings of
just `5 45` (needed for preview diff).

Another approach here would be to suppress
the validation errors from binding integers
to partial parses, but I prefer the behavior
of plain integers to more closely match the
behavior of suffixed integers, and I do like
having the freedom to set an invalid value
just in case that ever comes up (invalid
values might mess with future bias, because
the current test of whether one decoded-time
is earlier than another currently assumes
each part has a value in the valid range - I
want to wait and see if that actually proves
to be a problem in practice, once I actually
see a use for invalid values).