.zshrc: set TZ=utc on all my devices
I've decided to try going back to using local
time at the desktop level, and to see if the
automatic timezone stuff on laptops finally
works anywhere near as well as on phones.

But I still want UTC time by default in CLI
programs, for now.

In the long run, I'm not sure what I'll do.

My ideal is simply two clocks instead of one
in all of the standard places where time is
shown (but that seems to be largely unavailable
as a feature out-of-the-box in modern operating
systems), slash to invest more in my ability to
mentally convert between UTC and local time.

But for now I guess I'm settling into this
middle ground where my normie UX stuff is set
to local time and my techie UX stuff uses UTC.