.emacs.el: histdir: keep latest local entry first
Basically the only downside of fully live-synced
command history is that I have muscle memory and
habitualized cognition for "run the last (local)
command again", so sometimes I'll switch to a
shell/REPL and I think/intend/move to re-run the
last command from that instance (which is in the
scrollback, right above the command prompt), but
another entry has sync'ed in as more recent.

The full/proper solution would let us configure
(per-buffer!) how many local commands to store
before the shared history. This MVP commit only
stores the one most recent command "before" the
shared history.

If the latest local command is also the latest
command globally (and thus is the front of the
synced/shared history), Up from a fresh input
prompt shows the latest command twice. The
full/proper solution would hide that quirk
when deduplication is in use, at least if the
local recent commands are also the most recent
entries in the shared/synced history.

For now, the latest local history entry is not
accessible by hitting Down from older/shared
entries - it's accessible through hitting Up
and commands based on consult-history.
This is basically bug-flavored, but it's "good
enough" for me for now - this blemish doesn't
break anything in my unconditional "fast path"
muscle memory / mental habits, and doesn't
detract from the main goal/benefit of this
feature.