.emacs.el: "  j": done + schedule repetition
Just as " j" jogs a task down to "later" (actually
to the current time, but for unscheduled tasks the
datetime is just for sorting), prefixed " j" (so,
for example, "  j") jogs a scheduled task down to
its next repetition.

The difference from "zj" is that "zj" moves along
with the file to its new position (intended for
exploring/viewing use-cases), while "  j" just
renames the file without following it, like " j"
(intended for "done / will do later" use-cases).