.emacs.el: task-create: go to created task in list
Going to the last file in the list was right when
I didn't yet have scheduled tasks, so no task had
a datetime newer than the newly created task, and
so going to the end of the task list implicitly
went to the new task.

Now that I regularly have tasks scheduled for the
future in my task list, it's time to add the tiny
wrinkle of actually moving to the newly created
task in the list, wherever it is.