.emacs.el: try "  g" for date-then-title task create
I had been intending to use " a" for this, but I
formed that thought before I had prefix argument
bindings in space-map (i.o.w., before "  g" was
possible as an ergonomic variant of " g").

I may end up using " a" for it if I use it a lot,
since one less keypress is something I feel
enough to be bothered/drained by in common tasks.
But "  g" seems worth a try, and takes a little
less code change.